Every new customer now starts with 1000€ on his checkingAccount

+++ src/main/java/de/marcoedenhofer/edenbank/application/bankaccountservice/IBankAccountService.java
@@ -8,7 +8,7 @@ public interface IBankAccountService {
     void createCheckingAccountForCustomerAccount(CustomerAccount customerAccount);
     void createSavingsAccountForCustomerAccount(CustomerAccount customerAccount);
     void createFixedDepositAccountForCustomerAccount(CustomerAccount customerAccount);
-    void createCheckingAccountWithFixedBudged(CustomerAccount customerAccount, int budget);
+    void createCheckingAccountWithFixedBudged(CustomerAccount customerAccount, long budget);
     BankAccount loadBankAccountWithId(long bankAccountId) throws BankAccountNotFoundException;
     BankAccount loadBankAccountWithIban(String iban) throws BankAccountNotFoundException;
     void archiveBankAccount(BankAccount bankAccount);
