+++ src/main/java/de/marcoedenhofer/edenbank/application/interestservice/InterestService.java
@@ -8,7 +8,6 @@ import de.marcoedenhofer.edenbank.persistence.entities.BankAccount;
 import de.marcoedenhofer.edenbank.persistence.entities.FixedDepositAccount;
 import de.marcoedenhofer.edenbank.persistence.entities.SavingsAccount;
 import de.marcoedenhofer.edenbank.persistence.repositories.IBankAccountRepository;
-import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.context.annotation.Scope;
 import org.springframework.scheduling.annotation.Scheduled;
 import org.springframework.stereotype.Service;
@@ -86,7 +85,7 @@ public class InterestService implements IInterestService {
         }
     }
 
-    private double computeInterestPayment(int balance, float interest) {
+    private double computeInterestPayment(long balance, float interest) {
         return balance * (interest/100.0) / 100.0;
     }
 }
